Issue #75 - Add support for cloning of docs and copying of sheets
The way this PR is intended to work is the following:

Before fetching bug data for a given week, clone into a new document.
Fetch all bug data.
Copy the sheet into the original document.
Delete the cloned document.

The way this works assumes you have a stable spreadsheetID in config.json -
if you don't, it will write a spreadsheetID of the first created document
for you.

This should reduce the number of history entries and allow us to run in automation.
(in theory, we need to test this)